Remove int type (without size) from FOAST, PAST, ITIR

+++ tests/next_tests/unit_tests/type_system_tests/test_type_translation.py
@@ -65,7 +65,7 @@ def test_invalid_scalar_kind():
             typing.Tuple[int, float],
             ts.TupleType(
                 types=[
-                    ts.ScalarType(kind=ts.ScalarKind.INT),
+                    ts.ScalarType(kind=ts.ScalarKind.INT64),
                     ts.ScalarType(kind=ts.ScalarKind.FLOAT64),
                 ]
             ),
@@ -77,7 +77,7 @@ def test_invalid_scalar_kind():
                     ts.ScalarType(kind=ts.ScalarKind.BOOL),
                     ts.TupleType(
                         types=[
-                            ts.ScalarType(kind=ts.ScalarKind.INT),
+                            ts.ScalarType(kind=ts.ScalarKind.INT64),
                             ts.ScalarType(kind=ts.ScalarKind.FLOAT64),
                         ]
                     ),
@@ -102,10 +102,10 @@ def test_invalid_scalar_kind():
             ts.FunctionType(
                 args=[
                     ts.ScalarType(kind=ts.ScalarKind.FLOAT64),
-                    ts.ScalarType(kind=ts.ScalarKind.INT),
+                    ts.ScalarType(kind=ts.ScalarKind.INT64),
                 ],
                 kwargs={},
-                returns=ts.ScalarType(kind=ts.ScalarKind.INT),
+                returns=ts.ScalarType(kind=ts.ScalarKind.INT64),
             ),
         ),
         (typing.ForwardRef("float"), ts.ScalarType(kind=ts.ScalarKind.FLOAT64)),
